requirements.txt
Django==1.11.20
django-braces==1.11.0
django-crispy-forms==1.7.2
psycopg2==2.8.2
pytz==2019.1
